+++ config/default.example.js
@@ -1,0 +1,17 @@
+const config = {
+    
+      app: {
+        port: 3000
+      },
+      // 数据库配置
+      db: {
+        DATABASE: 'pieces_db',
+        USERNAME: 'root',
+        PASSWORD: '123456',
+        PORT: '3306',
+        HOST: 'localhost'
+      }
+    }
+    
+    module.exports = config
+