+++ src/server/controllers/index.js
@@ -1,0 +1,7 @@
+/**
+ * 主页控制器
+ */
+
+module.exports = async ctx => {
+    await ctx.render('index.html');
+}