+++ src/server/models/tags.js
@@ -1,0 +1,18 @@
+module.exports = (sequelize, DataTypes) => {
+    let Tag = sequelize.define('tag', {
+        name: {
+            type: DataTypes.STRING,
+            allowNull: false
+        }
+    }, {
+        underscored: true
+    });
+
+    Tag.associate = models => {
+        Tag.belongsToMany(models.post, {
+            through: "PostTag"
+        })
+    };
+
+    return Tag;
+};