+++ webpack.config.js
@@ -19,7 +19,18 @@ module.exports = {
                         presets: ['env', 'react']
                     }
                 }
-            }
+            },
+            {
+                test: /\.css$/,
+                use: [
+                    {
+                        loader: "style-loader"
+                    }, {
+                        loader: "css-loader"
+                    }
+                ]
+            },
+            { test: /\.(png|woff|woff2|eot|ttf|svg)$/, loader: 'url-loader?limit=100000' }
         ]
     }
 };